Задание 2.43. Поменял оформление и исправил ошибки в задачах 3-5
